+++ src/Smartstore/Engine/SmartConfiguration.cs
@@ -1,8 +1,13 @@
-﻿namespace Smartstore.Engine
+﻿#nullable enable
+
+using System.Net;
+using Microsoft.AspNetCore.HttpOverrides;
+
+namespace Smartstore.Engine
 {
     public partial class SmartConfiguration
     {
-        public string ApplicationName { get; set; }
+        public string ApplicationName { get; set; } = default!;
         public bool EnableDynamicDiscovery { get; set; } = true;
         public string ThemesBasePath { get; set; } = "/Themes";
         public string ModulesBasePath { get; set; } = "/Modules";
@@ -47,7 +52,7 @@
         /// <summary>
         /// The default schema to use for database objects. Set <c>null</c> to use database default schema (e.g. "dbo").
         /// </summary>
-        public string DbDefaultSchema { get; set; }
+        public string? DbDefaultSchema { get; set; }
 
         /// <summary>
         /// Enabled second level database cache.
@@ -72,7 +77,7 @@
         /// Leave empty if path should be "App_Data/Tenants/{Tenant}/Media".
         /// In web farms, you should specify the same UNC or mapped network drive on each server.
         /// </summary>
-        public string MediaStoragePath { get; set; }
+        public string? MediaStoragePath { get; set; } = default!;
 
         /// <summary>
         /// Public base path to the media storage used to generate URLs for output HTML.
@@ -89,7 +94,7 @@
         /// <summary>
         /// List of module names to ignore during app installation.
         /// </summary>
-        public string[] IgnoredModules { get; set; }
+        public string[]? IgnoredModules { get; set; }
 
         /// <summary>
         /// Task Scheduler poll interval in minutes (1 recommended)
@@ -100,13 +105,13 @@
         /// Set this to your store's task scheduler base url
         /// when the automatic url resolution fails (e.g. 'http://www.mystore.com')
         /// </summary>
-        public string TaskSchedulerBaseUrl { get; set; }
+        public string? TaskSchedulerBaseUrl { get; set; }
 
         /// <summary>
         /// Set this to your store's PDF engine base url
         /// when the automatic url resolution fails (e.g. 'http://www.mystore.com')
         /// </summary>
-        public string PdfEngineBaseUrl { get; set; }
+        public string? PdfEngineBaseUrl { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating whether to store TempData in a cookie.
@@ -117,7 +122,7 @@
         /// <summary>
         /// Use with care!
         /// </summary>
-        public string EngineType { get; set; } = typeof(SmartEngine).FullName;
+        public string? EngineType { get; set; } = typeof(SmartEngine).FullName;
 
         /// <summary>
         /// Gets or sets a value indicating whether to display exception
@@ -138,11 +143,11 @@
         public long? MemoryCacheSizeLimit { get; set; }
 
         /// <summary>
-        ///  Gets or sets the maximum allowed size of any request body in bytes. When set
-        ///  to null, the maximum request body size is unlimited. This limit has no effect
-        ///  on upgraded connections which are always unlimited. This can be overridden per-request
-        ///  via <see cref="Microsoft.AspNetCore.Http.Features.IHttpMaxRequestBodySizeFeature"/>.
-        ///  Defaults to 30,000,000 bytes, which is approximately 28.6MB.
+        /// Gets or sets the maximum allowed size of any request body in bytes. When set
+        /// to null, the maximum request body size is unlimited. This limit has no effect
+        /// on upgraded connections which are always unlimited. This can be overridden per-request
+        /// via <see cref="Microsoft.AspNetCore.Http.Features.IHttpMaxRequestBodySizeFeature"/>.
+        /// Defaults to 30,000,000 bytes, which is approximately 28.6MB.
         /// </summary>
         public long? MaxRequestBodySize { get; set; }
 
@@ -151,13 +156,79 @@
         /// </summary>
         public int SmtpServerTimeout { get; set; } = 5000;
 
+        /// <summary>
+        /// Gets or sets Google configuration options. 
+        /// </summary>
         public GoogleConfiguration Google { get; set; } = new();
+
+        /// <summary>
+        /// Gets or sets reverse proxy configuration options. 
+        /// </summary>
+        public ProxyConfiguration ReverseProxy { get; set; } = new();
 
         public class GoogleConfiguration
         {
-            public string RecaptchaWidgetUrl { get; set; }
-            public string RecaptchaVerifyUrl { get; set; }
-            public string MapsUrl { get; set; }
+            public string RecaptchaWidgetUrl { get; set; } = "https://www.google.com/recaptcha/api.js";
+            public string RecaptchaVerifyUrl { get; set; } = "https://www.google.com/recaptcha/api/siteverify";
+            public string MapsUrl { get; set; } = "https://maps.google.com/maps?hl={0}&q={1}";
+        }
+
+        public class ProxyConfiguration
+        {
+            /// <summary>
+            /// <c>true</c> enables the <see cref="ForwardedHeadersMiddleware"/>.
+            /// </summary>
+            public bool Enabled { get; set; }
+
+            /// <summary>
+            /// Processes the X-Forwarded-For header, which identifies the originating IP address of the client.
+            /// </summary>
+            public bool ForwardForHeader { get; set; } = true;
+
+            /// <summary>
+            /// Processes the X-Forwarded-Proto header, which identifies the protocol (HTTP or HTTPS) the client used to connect.
+            /// </summary>
+            public bool ForwardProtoHeader { get; set; } = true;
+
+            /// <summary>
+            /// Processes the X-Forwarded-Host header, which identifies the original host requested by the client.
+            /// </summary>
+            public bool ForwardHostHeader { get; set; } = true;
+
+            /// <summary>
+            /// Gets or sets the header used to retrieve the originating client IP. Default: "X-Forwarded-For".
+            /// </summary>
+            public string? ForwardedForHeaderName { get; set; }
+
+            /// <summary>
+            /// Gets or sets the header used to retrieve the value for the originating scheme (HTTP/HTTPS). Default: "X-Forwarded-Proto".
+            /// </summary>
+            public string? ForwardedProtoHeaderName { get; set; }
+
+            /// <summary>
+            ///  Gets or sets the header used to retrieve the original value of the Host header field. Default: "X-Forwarded-Host".
+            /// </summary>
+            public string? ForwardedHostHeaderName { get; set; }
+
+            /// <summary>
+            /// Addresses of known proxies to accept forwarded headers from.
+            /// </summary>
+            public string[]? KnownProxies { get; set; }
+
+            /// <summary>
+            /// The allowed values from X-Forwarded-Host. If the list is empty then all hosts are allowed.
+            /// Failing to restrict these values may allow an attacker to spoof links generated by your service.
+            /// </summary>
+            /// <remarks>
+            /// <list type="bullet">
+            /// <item><description>Port numbers must be excluded.</description></item>
+            /// <item><description>A top level wildcard "*" allows all non-empty hosts.</description></item>
+            /// <item><description>Subdomain wildcards are permitted. E.g. "*.example.com" matches subdomains like foo.example.com, but not the parent domain example.com.</description></item>
+            /// <item><description>Unicode host names are allowed but will be converted to punycode for matching.</description></item>
+            /// <item><description>IPv6 addresses must include their bounding brackets and be in their normalized form.</description></item>
+            /// </list>
+            /// </remarks>
+            public string[]? AllowedHosts { get; set; }
         }
     }
 }