+++ src/Smartstore/Net/ConnectionInfoExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System.Net;
-using System.Net.Sockets;
 using Microsoft.AspNetCore.Http;
 
 namespace Smartstore
@@ -13,23 +12,27 @@ namespace Smartstore
         {
             Guard.NotNull(connection);
 
-            var remoteAddress = connection.RemoteIpAddress;
-            if (remoteAddress == null)
+            if (connection.RemoteIpAddress != null)
             {
-                return false;
-            }
-            
-            // If the RemoteAddress is an IPv6 address, the method converts
-            // it to an IPv4-mapped address using the MapToIPv4 method and
-            // checks if that address is a loopback address as well.
-            var isLocalAddress = IPAddress.IsLoopback(remoteAddress);
-            if (!isLocalAddress && remoteAddress.AddressFamily == AddressFamily.InterNetworkV6)
-            {
-                //return remoteAddress.Equals(localAddress);
-                isLocalAddress = IPAddress.IsLoopback(remoteAddress.MapToIPv4());
+                if (connection.LocalIpAddress != null)
+                {
+                    // When both local and remote IP are equal, we are local.
+                    return connection.RemoteIpAddress.Equals(connection.LocalIpAddress);
+                }
+                else
+                {
+                    // If for some reason local IP is null, we are local when remote IP is a loopback.
+                    return IPAddress.IsLoopback(connection.RemoteIpAddress);
+                }
             }
 
-            return isLocalAddress;
+            // For in memory TestServer or when dealing with default connection info  
+            if (connection.RemoteIpAddress == null && connection.LocalIpAddress == null)
+            {
+                return true;
+            }
+
+            return false;
         }
     }
 }